finish up functionality for code
