+++ src/action/index.js
@@ -1,4 +1,4 @@
-import { CREATE_BOOK, REMOVE_BOOK } from './..action-shade';
+import { CREATE_BOOK, REMOVE_BOOK } from '../action-shade';
 
 export const createBook = book => ({
   type: CREATE_BOOK,
